feat: add persistAndFlush() and persistLater() methods
EM.persist() takes second bool parameter to decide whether we want to flush
immediately, or we will do it later. This can be a bit hard to understand
when you are not familiar with this concept. User can now choose from
dedicated methods.

It is also more convenient as EM.persistLater() is synchronous, so you can
freely use it in sync methods and handle flushing at later execution stage.

This also moves cascading logic to UoW and simplifies it a bit.