fix: do not require PK when computing change set
When committing multiple un-managed entities, we do not have the PK, but
once we persist the entity, we update `originalEntityData` map, which was
not taken into account when computing latter change sets.

With this fix, change sets will be computed properly even in case we
persist multiple items that references same entity (previously it was
firing subsequent update queries with no new data).