fix: support self-referencing in many to one association
Previously when you wanted to persist an entity that referenced itself,
you would need to do this in 2 steps (persist the entity first without
referencing self, then again to update it with the reference).

Trying to do this in one step would fail with infinite loop. With this
fix, persisting the entity with self reference included will work as
expected, firing 2 consecutive queries automatically.