feat(schema): use knex in schema generator
SchemaGenerator now offers `createSchema` and `dropSchema` methods that by default return the necessary SQL, as well as with a `run` flag enabled it can run the SQL directly.

There is also `SchemaGenerator.execute()` that breaks the query into separate lines and executes it on current connection. This method is used under the hood when using the `run` flag.

BREAKING CHANGES:
`SchemaGenerator.generate()` is now async.