chore: hide knex's warning about sqlite FK failures
knex does not allow adding new columns with FK in sqlite, while it is possible to do. this comments out the code that was
causing the warning to appear, in future we should implement custom alter table for sqlite that will handle this (or ideally
fix this in knex directly).

https://github.com/knex/knex/issues/3351