Fix undefined variables in migration
Column names are capitals in MySQL 8.0.19 (MySQL Community Server - GPL). Without the aliases, the columns will be returned like COLUMN_NAME, DATA_TYPE, COLUMN_DEFAULT etc. and causing undefined column properties (for example line 95) and breaking the migration if the table structure is existent (creating the structure from scratch works fine, it is just the updating that is broken).