feat(sql): use collection snapshots to compute precise diff (#405)
When updating m:n collections, snapshots are now used to compute the exact different with the current state.
This means that we now execute only the queries that are needed (e.g. if we only append 1 item, we do not
clear the whole collection and insert everything as it was before, but rather fire 1 precise query for the
missing item).