feat(core): allow for the number zero as a primary key
There are some cases where an entity will have a primary key of 0.
Our use case is a table of static statuses that are pervasive
throughout our application. We add the rows when we generate the
schema, not through the ORM. We only want to map relationships
to those statuses with the ORM. Some of these statuses start with
the id of 0 versus 1. Previous to this fix, there were generic
falsy checks that would take a primary key with a value of 0 as
if the primary key was not set. This fix makes those checks
specific for both null and undefined values.

BREAKING CHANGE: entities that were using a primary key of 0 to
signify that the primary key was not defined will now interpret
the zero to be the actual primary key.

To migrate, no longer default the primary key to 0. Use null
and undefined to signify a primary key that is not defined.