fix(sql): simplify transactions handling with knex
TX handling was done wrong, causing the TX to be closed/rolled back twice. That results into a hidden error inside a knex debug log,
so it was hard to find, also the version of knex we used to initial integration was buggy and the original code was somehow required.
With recent knex versions, this works as expected.