feat(discovery): use both entity name and path as key in MetadataStorage
Decorators are using static MetadataStorage, now the keys includes also the path to entity, so technically it is possible to have multiple entities with the same name (although not in the same ORM context, but with multiple ORM instances).

This should also fix issues with HMR and "multiple property decorators used" validation error.