fix(postgres): keep `bigint`/`numeric` types as string
Previously we were converting numeric columns to floats, which could lead to loosing precision. This reverts it,
and keeps the value retrieved from driver as string to be consistent with other drivers.

One should use custom type to handle the conversion to number if needed. In future commits, there will be probably
new custom type added for decimal column, and one for bigint is already provided.

Closes #324