feat(ts-morph): use `.d.ts` files for ts-morph discovery (#616)
This allows to analyse `.d.ts` files instead of the actual entity source files. Also
instead of analysing the whole project based on `tsconfig.json`, we now use the list
of paths to actual entities (we know them from the decorators), changing the extension
from `.js` to `.d.ts`.

BREAKING CHANGE:
Removed `discovery.tsConfigPath` option as it is no longer needed, it was used only
for `TsMorphMetadataProvider`, when the `entitiesDirsTs` were not explicitly provided.
In v4, this is no longer needed, as ts-morph discovery will use `d.ts` files instead,
that should be located next to the compiled entities.