feat(core): support globs in `entities`
`entitiesDirs` and `entitiesDirsTs` were removed in favour of `entities` and `entitiesTs`,
`entities` will be used as a default for `entitiesTs` (that is used when we detect `ts-node`).

`entities` can now contain mixture of paths to directories, globs pointing to entities,
or references to the entities or instances of `EntitySchema`. Negative globs are also
supported.

This basically means that all you need to change is renaming `entitiesDirs` to `entities`.

```typescript
MikroORM.init({
  entities: ['dist/**/entities', 'dist/**/*.entity.js', FooBar, FooBaz],
  entitiesTs: ['src/**/entities', 'src/**/*.entity.ts', FooBar, FooBaz],
});
```

BREAKING CHANGE:
`entitiesDirs` and `entitiesDirsTs` are removed in favour of `entities` and `entitiesTs`.

Closes #605