feat(core): execute hooks via `EventManager`
Lifecycle hooks are now executed the same way as `EntitySubscriber`s, this means that they
will also get the `EventArgs` in the parameters.

`EventArgs` now also optionally contain the `ChangeSet` object, so it is possible to get
the information about what fields actually changed or access the original entity data.

Closes #622