feat(core): add support for flush events
There is a special kind of events executed during the commit phase (flush operation).
They are executed before, during and after the flush, and they are not bound to any
entity in particular.

- `beforeFlush` is executed before change sets are computed, this is the only
  event where it is safe to persist new entities.
- `onFlush` is executed after the change sets are computed.
- `afterFlush` is executed as the last step just before the `flush` call resolves.
  it will be executed even if there are no changes to be flushed.

Flush event args will not contain any entity instance, as they are entity agnostic.
They do contain additional reference to the `UnitOfWork` instance.

Closes #637