fix(core): allow having same property name as entity name
Previously `QueryBuilder` was blindly expecting that any populate request for a property
that name collides with an entity name means we want to do M:N population.

This fixes the check by verifying the `meta.pivotTable` flag.

Also new method on `MetadataStorage` that returns possibly undefined is added to improve
typings in internals.

Closes #655