feat(core): add `driver.nativeInsertMany()` method
This method can be used for bulk inserts. It works in all existing drivers,
but has some limitations in SQLite and MySQL/MariaDB. Those drivers do not
return all the just created PKs.

MySQL/MariaDB will return PK of the first entity in the batch. If we have
`innodb_autoinc_lock_mode` set to 0/1 it is possible to compute the following
PKs simply by incrementing the first PK value.
(see https://stackoverflow.com/a/16592867/3665878)

SQLite on the other hand will return only the last PK.

Postgres and MongoDB both return all the PKs in the `res.rows` array.

Related: #442