fix: use lazy registered EventSubscribers
Before for every new context an EM was created with a new EventManager.
This EventManager would only use EventSubscribers that were annotated or explicitly defined in the
config.
Now the EventManager from the global EM is reused in every context specific EM.
Thereby lazy registeren EventSubscribers are also listening to events.

Closes #711