feat(core): refactor merging to allow querying by custom type (#800)
This changes internal mechanism of `em.merge()`. Originally it was using
`EntityAssigner` inside `em.merge()` together with `UoW.merge()`, which
was called twice. Because of this, the custom types were processed multiple
times, and the value passed to conversion method was often already converted.

Now we use only the `Hydrator` for this. As a sideproduct of those changes,
we can now query by custom types.

```typescript
class UuidBinaryType extends Type<string, Buffer> {

  convertToDatabaseValue(value: string): Buffer {
    return Buffer.from(parse(value));
  }

  convertToJSValue(value: Buffer): string {
    return stringify(value);
  }

  getColumnType(): string {
    return 'binary(16)';
  }

}

@Entity()
class A {

  @PrimaryKey({ type: UuidBinaryType })
  id: string = v4();

}

const a = await em.find(A, { id: '...' });
```

Closes #739