feat(entity-generator): do not use ts-morph
Again, not much need for ts-morph (a bit more than with migrations tho),
and it brought the full TS as a dependency.

With this change, the CLI package will be no longer having runtime
dependency on ts-morph and therefore TS, so it will be fine to install
it direct dependency (not a dev one).