perf(core): do not use `em.merge()` internally
Every managed entity did pass thru `em.merge()` originally. While it was a good invariant
for managed entities, it turns out the method was overcomplicated and had a serious performance
impact on the persistence.

With this change, UoW won't use `em.merge()` during flush operation. Instead, entity factory is
used directly to create new instance and new method - `uow.registerManaged()` is used to add it
to the identity map.

The method `em.merge()` is still present and used is some parts of the ORM like in `assign` helper,
it is also still a valid goto method for loading entities from cache.