perf(core): use batch inserts in UoW (postgres & mongodb)
UoW will now batch all inserts statements if the driver supports returning all of their
PKs back (so postgres and mongodb currently).

This also changes how `Collection.contains()` was implemented - it did a linear check
on all values, comparing by the PK for persisted entities. This took extreme amount of
time with huge collections, just to add to a collection 1000 times, it was as slow as
1s for this, without persisting anything.

Now the collection items are stored as `Set`, and `contains` method does a simple identity
lookup (`set.has(entity)`). Another issue that slowed it down was the array like index support,
where we are able to use `coll[0]` to access the first item in the collection. This was
implemented as a re-assigning of all items to the collection object, which was extremely slow
for large collections. Now when we add to a collection, we only append the last index instead
of reassigning everything.

Closes #732