fix(migrations): respect custom file names when running by name
When running migrations to specific version, we can either provide the full migration name,
like `Migration20191013214813`, or just the timestamp, e.g. `20191013214813`. But this was
not respecting the custom file name callback, and blindly expected the `Migration` prefix.

With this fix, the runner will check if the provided name is just the timestamp (a numeric
string of 14 digits, possibly with an extension), if so, it will use the `fileName` callback
to construct the full migration name for you.

Closes #883