fix(core): ensure custom types are comparable
When we hydrate an entity, we convert the custom type to the JS runtime value.
This value is later converted back to the DB value and compared with the original
entity data snapshot. In some cases, this failed to recognize equal values as
the db returned the value in slightly different format (like `0.000` instead of `0`).

With this change, we also convert the hydrated value back to the DB type and update
the snapshot with this value to ensure it will be comparable.

Also the Buffer type was not compared correctly, which is now fixed.

Closes #864