perf(core): interpolate query parameters at ORM level
To get around limitations on query parameter limits, as well as to make it faster,
we now interpolate all query params manually, passing raw query to the connector.

This also allows to print the actual queries in the debug log. Previously output of
knex was used, but that was just an estimation, not a real query (e.g. dates were
presented as ISO strings, but stored as numbers/timestamps).

+ Mapping of DB results now also uses JIT compilation.

Related: #732