fix(core): validate the object passed to `em.persist()`
This catches errors that were caused by passing not discovered entity to `persist()` and uses more
readable error message with a suggestion if such entity name is known.

Previously it was failing  with errors like `TypeError: Cannot read property 'relations' of undefined`,
because the metadata were not present on the prototype.