fix(core): use `$and` for merging of multiple filter conditions
Previously multiple filter conditions were object-merged, which could result
in ignoring the left-hand side conds sometimes, especially when it comes to
merging group operators. This changes the way we merge them, so instead of
object-merging, we now craft an `$and` query with all the child conditions.

As part of this it was also needed to adjust how child conditions are being
created for select-in population. It should now also work with group operators,
which were previously ignored.

Closes #1776