refactor(core): return `Collection` instance from `LoadedCollection.get()/$`
Closes #2394

BREAKING CHANGE:
Previously those dynamically added getters returned the array copy of collection
items. In v5, we return the collection instance, which is also iterable and has
a `length` getter and indexed access support, so it mimics the array. To get the
array copy as before, call `getItems()` as with a regular collection.