feat(core): add support for concurrency checks
As opposed to version fields that are handled automatically, we can use
concurrency checks. They allow us to mark specific properties to be included
in the concurrency check, just like the version field was. But this time, we
will be responsible for updating the fields explicitly.

When we try to update such entity without changing one of the concurrency fields,
`OptimisticLockError` will be thrown. Same mechanism is then used to check whether
the update succeeded, and throw the same type of error when not.

```ts
@Entity()
export class ConcurrencyCheckUser {

  // all primary keys are by default part of the concurrency check
  @PrimaryKey({ length: 100 })
  firstName: string;

  // all primary keys are by default part of the concurrency check
  @PrimaryKey({ length: 100 })
  lastName: string;

  @Property({ concurrencyCheck: true })
  age: number;

  @Property({ nullable: true })
  other?: string;

}
```

Closes #1638