fix(core): do not override internal EM instance when forking EM
When running `em.transactional()`, all currently loaded entities are merged into the inner EM fork.
But when that happens, their internal EM instance is updated to the fork, and it stays like that
after you exit the transactional handler. So by calling the `loadItems()` (or any other method that
works with the internal EM instance, e.g. `wrap(e).init()`), we would try to use the inner fork,
that should no longer exists (and its transaction that is already finished).

Closes #2342