fix(core): issue early delete queries for recreating unique properties
Delete queries are normally issued at the very end of flush operation. When we
insert and delete an entity with a unique property value in the same unit of work,
we need to delete first to get around unique constrain violation. We now detect
those situations and remove such entities first.

Closes #2273