feat(core): validate required properties before flushing new entities
Closes #2176

BREAKING CHANGE:
Previously the validation took place in the database, so it worked only for SQL
drivers. Now we have runtime validation based on the entity metadata. This means
mongo users now need to use `nullable: true` on their optional properties, or
disable the validation globally via `validateRequired: false` in the ORM config.