feat(core): add custom table check constraint support for postgres
Adds a new `@Check` decorator for entities, allowing to set up custom check
constraints. At the moment, this is only for Postgres, as MySQL only supports
check constraints as of version 8, which is above the minimum test version.

While SQLite does in theory support check constraints already, tests broke due
to some bug in Knex, which is why the implementation was skipped for now.

This partially addresses #1711