fix(core): allow onUpdate access to other onUpdate results from earlier properties
Allow onUpdate to get access to previous entity values returned via onUpdate,
which can be useful for ensuring created and updated dates have the exact same time
or basing the value of one property off the value of another.

This already worked for onCreate, as onCreate directly updates the entity.
The onUpdate hook however stores its values in a map to ensure values are not updated
unless they actually need to be, which means onUpdate would always get property values
not accounting for prior onUpdate hooks.