fix(core): allow changing PK via UoW
The update queries now use the original PK before it was changed, resulting in
correct queries that use new PK only in data, but the old one in condition.

This commit also removes the `values` keyword usage in sqlite queries, as sqlite
apparently already supports tuple comparison without it, so we don't need special
handling for it.

Closes #3184