feat(core): automatically discover target embeddables and relationshipts
When we see a missing target entity or embeddable that is defined with a reference,
e.g. `@Embedded(() => Profile)` or `@ManyToOne(() => Author)`, we now try to discover
the target automatically. This means we no longer need to care about discovering
embeddables explicitly as long as we specify them correctly in owning entities.
This works recursively.