feat(core): automatically detect `src/dist/build` folders and adjust configuration (#3497)
If there is a `src` folder, the ORM will now automatically detect it and adjust the `seeder` and `migrations` configs to have both `path` and `pathTs` set based on that. This happens only if there are no values provided by user and the default folders don't exist.

For a fresh project, this means that if you omit the configuration, but your project has a `src` and `dist` folders existing, the ORM will automatically set the following:

- `migrations.path` to `./dist/migrations`
- `migrations.pathTs` to `./src/migrations`
- `seeder.path` to `./dist/seeders`
- `seeder.pathTs` to `./src/seeders`

In addition to this, the CLI no longer requires explicit configuration of `configPaths` in `package.json`, as long as your config is called `mikro-orm.config.js/ts` and is located in `src` or `dist/build` folder. `useTsNode` flag is still required.