feat(schema): try to infer runtime default values automatically
When a runtime default value is provided and the entity instance can be constructed without
providing any parameters, the default values will be inferred automatically.

We try to create two entity instances to detect the value is stable, and compare the values
by reference, this will discard things like `new Date()` or any object values in general.