perf(core): improve support for large collections
Previously `Collection.add/remove` were defined with a spread parameter, which could result in perf
degradation in very large collections. This change makes the first parameter accept also an array of
items, skipping unnecessary copying of arrays when hydarting the collection.

This is slightly breaking on type level, as it is no longer possible to do `coll.add(...items)` with
a generic array - it only works with a tuple. The solution is to do `coll.add(items)` instead, which
will type check and is also more performant (no array copies).

Closes #3572