fix: respect postgresql no timestamptz precision default (#3832)
See details in https://github.com/mikro-orm/mikro-orm/issues/3824. This
PR removes the default fallback to a precision of 0 so that for
postgresql the correct type for Date (timestamptz with no precision) is
generated, and it adds the fallback to 0 downstream in platform-specific
contexts.
(this PR is a duplicate of
https://github.com/mikro-orm/mikro-orm/pull/3826, reopened against the
`v6` branch)