fix(core): make `FilterQuery` strict again!
The inference fix from fcb17392 introduced an issue with `FilterQuery`
that started accepting any keys. This was because of the extension, which
was not a proper way to fix this. The original issue was in fact caused by
something else than the PK type - it was more about the usage of naked `T`
type in the `FilterQuery` union. We now define it as `T & {}` which helps
with getting the right inference priorities. The fix from fcb17392 is now
removed and the types are simplified (the `Query` generic is removed).

Related: fcb17392