fix(core): use types from `discovery.getMappedType` on runtime too
Previously, this option was used only to determine the database column type,
now it will be also used for the runtime type implementation, so equivalent
to doing `@Property({ type: MyType })`. This works only for types created
by the user, as most of the ORM mapped types are not relevant on runtime and
doing so for them would add a small performance penalty.

Closes #4042