fix(core): map returned values correctly after batch insert
When mapping the values from returning statement back to the entity, we mistakenly
used `res.row` for all the entities from the batch. This was rather hidden bug,
as primary keys are handled separately, so the use case was mainly things like
timestamps (which won't reproduce this issue).

But if the returning values were different for each row, e.g. some UUID (that is not PK),
this commit makes sure we map the correct rows to matching entities.