perf(core): improve result mapping and snapshotting
Previously we had to snapshot the entity after it was loaded from database
and hydrated, because the data could be in a different format (e.g. due to
joined strategy). This is now improved and the snapshotting is no longer
needed.

This has a nice side effect, custom types "to db" methods are no longer
triggered after entity gets loaded, only the "from db" ones are.