fix(core): rework JSON value processing (#4194)
Every driver behaves a bit differently when it comes to handling JSON
columns. While SQLite is not processing then anyhow, and MongoDB
supports them natively, all the others have rather quirky and not very
configurable JSON parsing and stringification implemented. It is crucial
to have the JSON values properly parsed, as well as normalized in the
entity snapshot, so we can correctly detect updates.

This PR makes the JSON parsing fail-safe, returning the value directly
if it is not a valid JSON string, and ensures the entity data are in the
right shape.

Closes #4193