feat(core): deprecate `persist/flush/remove` methods from `EntityRepository`
Both persist and flush are tied to the EM instance, we have the repository shortcuts, but they
feel confusing, as they might give users false sense that they do something only with that given
entity type. Maybe we should remove them and make the em getter public instead, to preserve easy
access to the functionality.

In preparation for removal of those methods in v6, all the following `EntityRepository` methods
are now deprecated:

- persist
- persistAndFlush
- remove
- removeAndFlush
- flush

Additionally, there is a new public `repo.getEntityManager()` method available.

Related: #3989