fix(core): fix extra updates caused by bigint type
Mapped types have a `ensureComparable` option that allows them to ensure
the mapped value will be comparable, this is used to normalize things like
JSON or bigint shapes.

While the mechanism itself was working, we were calling `registerManaged`
twice, and only the first call did hydration and therefore had correct
values. In some cases, the second call did override the original entity
data back to the numeric value (which might be returned from mysql/sqlite).

This removes the second call, as it was clearly not necessary, this is
already handled from inside  the `factory.create()` with `merge: true`.

Closes #4249