fix(core): rework `Collection` initialization to use `em.populate()` (#4571)
This unifies the population mechanism for to-many relations by using the
`em.populate()` internally.

Previously, the `Collection.init` (and methods using it, e.g.
`loadItems`) were using custom implementation, which in some cases
resulting in different results as opposed to `em.populate()`. With this
PR, the `init` method is also using `em.populate`, yielding the same
results as if you would populate the relation.

Closes #4464