fix(core): rework JSON property hydration
The JSON properties are no longer mapped automatically on the driver level. The mapping from JSON strings is now implemented in the result mapping layer.

The motivation behind this change was to support property mapping of numeric values, and finally put an end to the JSON mapping bugs, that were mostly originating from small differences and bugs in how the underlying libraries that were trying to do it on their own.

Closes #4476