fix(core): ensure propagation and change-tracking works with `useDefineForClassFields` (#4730)
Propagation failed when `useDefineForClassFields` TypeScript compiler
flag was enabled (which is true when targeting `ES2022` or higher).

This PR ensures the propagation works regardless of the flag, by
ensuring the getters are set up when registering entity as managed (and
on few other places) via the new `EntityHelper.ensurePropagation()`
method.

Moreover, this PR also enables the `useDefineForClassFields` flag, and
since v6 is targeting es2022, we get no transpilation for class
properties thanks to this.

Closes #4216