feat(core): allow setting default `schema` on `EntityManager` (#4717)
Hi,

I want to start with thanking you all for this great ORM that is so
actively maintained and offer so many great features. One thing that
really catches my eye is the per request (shell)EntityManager that I
think will make this ORM a winner when it comes to multi tenancy. Almost
all of the other big ORM libraries are based on a global concept.

My aim with this PR is to make a shoot at implementing schema based
multi tenancy support with help of the per request EntityManager. With
this PR it's possible to set a schema when forking the EntityManager.
That schema will be applied as a default schema if no other schema is
present. It makes use of the schema options introduced in:
https://github.com/mikro-orm/mikro-orm/pull/2296