fix(core): make `em.create` strictly typed for relations too
When using `em.create()`, we can create the entity graph in one step, including relations. This commit makes the typing strict for the relations too, requiring their mandatory props to be present

There is one common issue with this - the `em.create` automatically propagates things to inverse sides, in other words, doing this is completely fine, as the author will be propagated to the new books:

```ts
em.create(Author, {
  name: '...',
  books: [{ title: 'foo' }], // Book.author is required, but will be propagated from the upper scope
});

```

To get around it, the child properties matching their parent type are automatically considered as optional - in other words, since we are processing contents of `Author.books`, we will make all properties of type `Author` in the `Author.books` as optional. It's not perfect, as it will mark every `Author` property as optional, not just the one that is an inverse of `Author.books`, but I feel like it's the tradeoff we need to make.

Closes #4748