fix(entity-generator): ambiguous FK generation
The process of generating relations is heavily altered
to output foreign key relations with property names based on the foreign key constrain name,
except in cases where it is safe to use a specific column name instead.

When composite foreign keys are outputted, all fields are included with "fieldNames",
unless all names involved at the same as the join columns.

In general, if a metadata property is an array, that array will be outputted correctly.

If a column is part of a foreign key (be it a composite one or ambiguous single one),
no "plain" column property is generated, unless all FKs are null-able, while the column is not.
Access to the property is through the foreign key properties instead.

Made the order of indexes and FKs in the information schema select explicit,
based on the name and definition order of the columns,
thus ensuring a consistent output.

Tweaked the OneToOne relation detection to also work for composite primary keys.
If the entire PK is covered by exactly one ManyToOne relation,
it now counts at OneToOne regardless of the number of columns involved.

Fixes #4898.