fix(entity-generator): nullable foreign keys and refactors
Refactored the foreign key processing into a separate private method.

Synced MariaDbSchemaHelper with MySqlSchemaHelper.

Removed the "isReferenced" property from EntityMetadata
in favor of a WeakSet in the entity generator itself.

Adjusted the OneToOne primary detection to work properly,
even when the target is a composite FK.

Fixed index property detection.
For composite indexes that are not in use by a FK,
attempt to map all columns of the index to properties.
If that is not possible, fallback to using an "expression".
Note that as a consequence of this change,
options of unique indexes are allowed to have an expression.
The SchemaHelper is adjusted to support this.

If a composite foreign key has multiple nullable columns,
properties for the individual nullable columns are added,
in addition to the property for the foreign key itself.
This allows the generated entity to represent
rows where only some of the columns are set,
and therefore, the FK constraint is not enforced.