feat(entity-generator): scalarFieldsForCompositeRelations config option
Added option to the entity generator called "scalarPropertiesForRelations".
Used to control how scalar properties for composite relations are generated.
Default value of "never" preserved BC and forces all FKs to be fulfilled,
value "always" renders all scalar properties of FKs always,
suitable mostly for applications that intend to sometimes disable FK checks
and value "smart" renders scalar properties of composite FKs only if
the FK can potentially be violated due to only some columns being set.

Adjusted the naming algorithm to fallback to removing the "reference column name"
when removing the full reference column name does not result in a change.
This would remove the "_id" suffix in naming conventions where
both sides of the relation are using the same column name,
and that name happens to end in "_id"
(or whatever a custom naming strategy would dictate that the reference column name is).

Fixed generation of indexes annotations for plain columns.