fix(entity-generator): prop order and scalar PK
Adjusted prop order, such that scalar props follow the FK prop named after them.

Scalar primary columns are not marked as such if they are not to be persisted.

Added handling (and a test for) an edge case where the FK constraint name could
in theory match the name of a column in the same table,
while also not being the only reference to that table.
In such cases, "fk_" is added as prefix, potentially multiple times,
until a non-existent column is hit.

Added tests for more setting combinations in GH4898.test.ts.
Specifically, for bidirectionalRelations and identifiedReferences.
esmImports is intentionally left out due to its smaller impact (not worth the time).